src/backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.4.26
cffi==1.17.1
channels==4.0.0
channels_redis==4.2.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.0.0
Django==5.1.5
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
selenium==4.32.0
service-identity==24.2.0
setuptools==76.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zope.interface==7.2